Documented a 4 more files (5/11)

+++ player.gs
@@ -19,12 +19,51 @@ class Player {
     }
     this.loadUnits()
     this.loadMoney()
-    var refund = unitType.buyFunction(1, this, true)
+    const selection = SpreadsheetApp.getActiveSpreadsheet().getSelection()
+    const selectedCell = selection.getCurrentCell()
+    if (this.findUnit(getFieldColumn(this.number, selectedCell.getColumn()))) {
+      ui.alert("there is already a unit there")
+      return
+    }
+    var refund = unitType.buyFunction(this, true, selectedCell)
     if (refund == 1) {
       return
     }
     if (this.availibleUnits >= 1) {
-        unitType.buyFunction(1, this, false)
+        unitType.buyFunction(this, false, selectedCell)
+        this.availibleUnits -= 1
+        this.shop.deleteItem(unitType)
+        this.saveMoney()
+        if (this.number == 1){
+          gameInfo.getRange("f2").setValue(this.availibleUnits)
+        } else {
+          gameInfo.getRange("g2").setValue(this.availibleUnits)
+        }
+        for (let unit of this.units) {
+          unit.update()
+        }
+        field.getRange("a1").getValue()
+        return
+      }
+    ui.alert("You already have the max units for this round")
+  }
+  buyForBots(unitType, column) {
+    let cell = this.activeSheet.getRange(1,column)
+    if (unitType.baseHealth == 0) {
+      return
+    }
+    this.loadUnits()
+    this.loadMoney()
+    if (this.findUnit(getFieldColumn(this.number, cell.getColumn()))) {
+      Logger.log(this.findUnit(getFieldColumn(this.number, cell.getColumn())).toArray())
+      return
+    }
+    var refund = unitType.buyFunction(this, true, cell)
+    if (refund == 1) {
+      return
+    }
+    if (this.availibleUnits >= 1) {
+        unitType.buyFunction(this, false, cell)
         this.availibleUnits -= 1
         this.shop.deleteItem(unitType)
         this.saveMoney()
@@ -35,7 +74,6 @@ class Player {
         }
         return
       }
-    ui.alert("You already have the max units for this round")
   }
   //
   createUnit(type, column) {
@@ -49,6 +87,11 @@ class Player {
     if (this.units.indexOf(unit) > - 1){
       this.units.splice(this.units.indexOf(unit), 1)
     }
+    field.getRange("a1").getValue()
+    for (let unit of this.units) {
+      unit.update()
+    }
+    field.getRange("a1").getValue()
   }
   //
   findUnit(column) {
